feat: Move Item Category parse into its own helper
