+++ src/renderer/stores/domain/item.ts
@@ -1,9 +1,7 @@
 import { makeAutoObservable, computed } from 'mobx';
 import { v4 as uuidv4 } from 'uuid';
-import Constants from '../../../helpers/constants';
 import { ItemData } from '../../../helpers/types';
-import { electronService } from '../../electron.service';
-const { logger } = electronService;
+import { getCategory } from '../../../helpers/item';
 
 type LootTableData = {
   id: string;
@@ -70,92 +68,6 @@ const getQuality = (data: ItemData) => {
 // Returns 0 if no quality is found
 const getRarity = (data: ItemData) => {
   return data.frameType < 1 || data.frameType > 3 ? 0 : data.frameType;
-};
-
-// Get the item category from all the itemData.
-// Lots of hardcoded stuff here, but it's the best way that we found.
-const getCategory = (item, subcategory = false) => {
-  // Read type for hybrid items too
-  let type = item.hybrid ? item.hybrid.baseTypeName : item.typeLine;
-  if (!type) return null;
-
-  const iconFileName = item.icon.replace(/^.*[\\/]/, '');
-
-  // Easiest cases
-  if (type === 'Expedition Logbook') return type;
-  if (type.includes('Blueprint')) return 'Blueprint';
-  if (type.includes('Contract'))
-    return Constants.items.names.heistQuestItems.includes(type) ? 'Quest Items' : 'Contract';
-
-  // Memories
-  for (const memory of Constants.items.names.memories) {
-    if (type.includes(memory)) return memory;
-  }
-
-  // Invitations
-  for (const invitation of Constants.items.names.invitations) {
-    if (type.includes(invitation)) return invitation;
-  }
-
-  // Misc basetypes that we know of
-  if (Constants.items.baseTypes.others[type]) {
-    const foundBaseType = Constants.items.baseTypes.others[type];
-    return !subcategory && Array.isArray(foundBaseType) ? foundBaseType[0] : foundBaseType;
-  }
-
-  // Items where we can guess the category from the frameType
-  // Order matters here, because some items need to be captured by this before we test them for the next categories
-  switch (item.frameType) {
-    // 4 = Gems
-    case 4:
-      const gemName = type.replace(/(Superior|Anomalous|Divergent|Phantasmal) /g, '');
-      if (gemName.endsWith('Support')) {
-        return 'Support Skill Gems';
-      } else {
-        return 'Active Skill Gems';
-      }
-    case 5:
-      if (type.startsWith('Captured Soul')) {
-        return 'Pantheon Soul';
-      } else {
-        return 'Labyrinth Items';
-      }
-    case 6:
-      return 'Divination Card';
-    case 7:
-      return 'Quest Items';
-    case 8:
-      return 'Prophecy';
-  }
-
-  // Maligaro's Map quest item has frameType 7, already detected above as a quest item
-  if (type.includes(' Map')) return 'Maps';
-  if (type.includes('Scarab')) return subcategory ? ['Map Fragments', 'Scarab'] : 'Map Fragments';
-  if (type.includes('Watchstone')) return 'Atlas Region Upgrade Item';
-  if (type.endsWith('Incubator')) return 'Incubator';
-  if (type.endsWith('Piece')) return 'Harbinger Item Piece';
-  if (item.icon.includes('BestiaryOrbFull')) return 'Captured Beast';
-
-  // Metamorph organs
-  if (Constants.items.names.metamorphSamples.includes(iconFileName.replace(/\..*$/, '')))
-    return 'Metamorph Sample';
-
-  // equipment - search by hardcoded basetype
-  // Remove quality if present
-  type = type.replace('Superior ', '');
-
-  // Non-magic equipment
-  if (item.frameType !== 1 && Constants.items.baseTypes.equipments[type])
-    return Constants.items.baseTypes.equipments[type];
-
-  // Magic equipment - typeline is polluted by prefixes $%&*#^@!!!
-  for (const baseTypeName in Constants.items.baseTypes.equipments) {
-    if (type.includes(baseTypeName)) return Constants.items.baseTypes.equipments[baseTypeName];
-  }
-
-  logger.error(`No category found for item ${item.id || '(no id)'}! JSON follows:`);
-  logger.info(item);
-  return null;
 };
 
 // Get the list of influences from itemData
